requirements.txt
asgiref==3.8.1
attrs==25.1.0
backports.zoneinfo==0.2.1
boto3==1.37.23
botocore==1.37.23
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
Django==4.2.19
django-allauth==65.4.1
django-cors-headers==4.4.0
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
idna==3.10
importlib_resources==6.4.5
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
numpy==1.24.4
oauth2_provider==0.0
oauthlib==3.2.2
pandas==2.0.3
pillow==10.4.0
pillow-avif-plugin==1.5.1
pkgutil_resolve_name==1.3.10
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.1
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
zipp==3.20.2
